Verificando de item

+++ src/items/items.controller.ts
@@ -1,7 +1,7 @@
 // src/items/items.controller.ts
 
 // NestJS
-import { Controller, Get, Post, Put, Delete, Param, Body, ParseIntPipe } from '@nestjs/common';
+import { Controller, Get, Post, Put, Delete, Param, Body, ParseIntPipe, BadRequestException } from '@nestjs/common';
 
 // Servicios
 import { ItemsService } from './items.service';
@@ -39,12 +39,42 @@ export class ItemsController {
     // Metodo para actualizar un item
     @Put(':id')
     async update(@Param('id', ParseIntPipe) id: number, @Body() updateItemDto: UpdateItemDto): Promise<Item> {
+        // Validar que el id del item y el id del DTO sean proporcionados
+        if (!id || isNaN(id)) throw new BadRequestException('Id es requerido.');
+        if (!updateItemDto.id) throw new BadRequestException('Id es requerido.');
+
+        // Validar que el id del item a actualizar sea el mismo que el id del item en el DTO
+        if (id !== updateItemDto.id) throw new BadRequestException('Id no coincide.');
+
+        // Verificar que el item exista
+        const item = await this.itemsService.findOne(id);
+
+        if (!item) throw new BadRequestException('Item no encontrado.');
+
         return await this.itemsService.update(id, updateItemDto);
     }
 
     // Metodo para eliminar un item
     @Delete(':id')
     async delete(@Param('id', ParseIntPipe) id: number): Promise<Item> {
+        // Validar que el id sea proporcionado
+        if (!id) throw new BadRequestException('Id es requerido.');
+
         return await this.itemsService.delete(id);
     }
+
+    // Metodo para verificar que un item existe
+    // verifyItem(id: number, itemDTO: UpdateItemDto | CreateItemDto): void {
+    //     // Validar que el id del item y el id del DTO sean proporcionados
+    //     if (!id || isNaN(id)) throw new BadRequestException('Id es requerido.');
+    //     if (!itemDTO.id) throw new BadRequestException('Id es requerido.');
+
+    //     // Validar que el id del item a actualizar sea el mismo que el id del item en el DTO
+    //     if (id !== itemDTO.id) throw new BadRequestException('Id no coincide.');
+
+    //     // Verificar que el item exista
+    //     const item = await this.itemsService.findOne(id);
+
+    //     if (!item) throw new BadRequestException('Item no encontrado.');
+    // }
 }
